Add programs to process NOAA GSOD data
This is a set of programs to download NOAA global surface summary of the
day data.
The first program will connect to the NOAA ftp server and download each
year's data, unpack it and upload it to HDFS.
The second one will convert the data to a single parquet file.
The third one does the same but will partition the parquet file by
year. Finally, the last program will compute the average surface
temperature per year and write the result to a single CSV file.